Move authorization into AuthController.
This was the only web route in the OAuthController, so I think it makes
sense to pull out into one of the web controllers.
