+++ app/Auth/Entities/UserEntity.php
@@ -4,8 +4,23 @@ namespace Northstar\Auth\Entities;
 
 use League\OAuth2\Server\Entities\UserEntityInterface;
 use League\OAuth2\Server\Entities\Traits\EntityTrait;
+use Illuminate\Contracts\Auth\Authenticatable;
 
 class UserEntity implements UserEntityInterface
 {
     use EntityTrait;
+
+    /**
+     * Make a new UserEntity from an Eloquent model.
+     *
+     * @param Authenticatable $user
+     * @return self
+     */
+    public static function fromModel(Authenticatable $user)
+    {
+        $entity = new UserEntity();
+        $entity->setIdentifier($user->getAuthIdentifier());
+
+        return $entity;
+    }
 }
